requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.10.22
psycopg2-binary==2.9.9
python-dotenv==1.0.1
ffmpeg-python==0.2.0